+++ changelog.md
@@ -6,6 +6,7 @@ whether or not to normalize by host population size in calculation of
 receive contact weights, equivalently, added `constant_transition_density`
 for receive transition weights (`constant_transition_density` should by
 default be `true`; can't imagine otherwise)
+- Add response loss event
 
 ## 3 February 2025
 Significant overhaul of weights calculations in order to accommodate correct